requirements_rag.txt
llama-index>=0.12
llama-index-vector-stores-chroma>=0.4
llama-index-embeddings-huggingface>=0.5
llama-index-llms-openai-like>=0.3
chromadb>=0.6
kss>=6.0
streamlit>=1.40
python-dotenv>=1.0
pandas>=2.0
pytest>=8.0
